requirements.txt
functions-framework
google-cloud-firestore
google-genai
python-dotenv
